Tidying up api
