+++ package.json
@@ -36,6 +36,7 @@
     "express": "^4.14.0",
     "jsonwebtoken": "^7.2.1",
     "mongodb": "^2.2.22",
+    "mongoose": "^4.8.1",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-router": "^3.0.2"
